Split crud.py (890 lines) into modular package
Convert utils/crud.py into a utils/crud/ package with entity-specific
submodules:
- _common.py: shared logging utilities (sanitization, log_db_operation)
- ml.py: ML analysis and annotation CRUD operations
- images.py: DataInstance CRUD, soft/hard deletion, deletion events
- reviews.py: ImageReview CRUD and project review status aggregation
- projects.py: users, projects, image classes, classifications, comments,
  project metadata, and API key operations
- __init__.py: re-exports all functions for full backward compatibility

All existing imports (import utils.crud as crud) continue to work
unchanged. The largest submodule (projects.py) is 398 lines, down from
the original 890-line monolith.

Updates one test to patch the logger at its new module path.
